% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMM_gate_keeper.R
\name{BMM.gate.keeper}
\alias{BMM.gate.keeper}
\title{BMM.gate.keeper}
\usage{
BMM.gate.keeper(data, sample.id, tracers)
}
\arguments{
\item{data}{Data frame. The dataset containing both source and target samples.}

\item{sample.id}{Character. The column name for sample identifiers.}

\item{tracers}{Character vector. The column names of the tracers used in the model.}
}
\description{
Checks whether specified tracer in a dataset contain zero values. If so, returns an error message listing the affected properties and corresponding sample names.
}
\author{
Thomas Chalaux-Clergue
}
