% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMM_pred.R
\name{BMM.pred}
\alias{BMM.pred}
\title{Predicted Source Contributions Extraction for Bayesian Mixing Models}
\usage{
BMM.pred(
  data,
  stats,
  sample.id,
  source,
  save.dir,
  note,
  fileEncoding = "latin1"
)
}
\arguments{
\item{data}{A data frame containing the predicted source contribution from a Bayesian Mixing model.}

\item{stats}{A character or a character vector specifying the summary statistics to extract for each source among: "Median" and/or "Mean".}

\item{sample.id}{A character specifying the column name in 'data' containing sample identifiers.}

\item{source}{A character specifying the column name in 'data' containing source identifiers.}

\item{save.dir}{Connection open for writing the test results data.frame. If "" save the file at working directory.}

\item{note}{A character string to add a note at the end of the file name (not set - default).}

\item{fileEncoding}{character string, if non-empty declares the encoding to be used on a file (not a connection) so the character data can be re-encoded as they are written, "latin1" (default).}
}
\value{
A data frame containing the predicted source contributions for each sample.
}
\description{
Predicted Source Contributions Extraction for Bayesian Mixing Models
}
\author{
Thomas Chalaux-Clergue
}
