% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMM_summary.R
\name{BMM.summary}
\alias{BMM.summary}
\title{Summarize the results of BMM predictions}
\usage{
BMM.summary(
  pred,
  sample.id,
  source,
  value,
  save.dir,
  note,
  fileEncoding = "latin1"
)
}
\arguments{
\item{pred}{A data frame containing the predicted values from a mixing model.}

\item{sample.id}{A character specifying the column name in 'pred' containing sample identifiers.}

\item{source}{A character specifying the column name in 'pred' containing source identifiers.}

\item{value}{A character specifying the column name in 'pred' containing predicted values.}

\item{save.dir}{Connection open for writing the test results data.frame. If "" save the file at working directory.}

\item{note}{A character string to add a note at the end of the file name (not set - default).}

\item{fileEncoding}{character string, if non-empty declares the encoding to be used on a file (not a connection) so the character data can be re-encoded as they are written, "latin1" (default).}
}
\value{
A data frame containing summary statistics for the predicted values.
}
\description{
Summarize the results of BMM predictions
}
\author{
Thomas Chalaux-Clergue
}
