% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinations.R
\name{combinations}
\alias{combinations}
\title{Combination generator}
\usage{
combinations(x)
}
\arguments{
\item{x}{a character vector of types/class.}
}
\value{
a character vector of combination.
}
\description{
Return the list of given type/class unique combinations.
}
\author{
Thomas Chalaux-Clergue
}
