% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRPS.R
\name{CRPS}
\alias{CRPS}
\title{Continuous Raking Probability Score}
\usage{
CRPS(
  obs,
  prev,
  path.to.prev,
  sample.name.obs,
  sample.name.prev,
  prev.source.name,
  prev.values,
  source.groups,
  mean.cal = FALSE,
  save.dir,
  note,
  fileEncoding = "latin1"
)
}
\arguments{
\item{obs}{a data.frame with the observed source group contributions.}

\item{prev}{a data.frame with previsions for each sample and for each source group.}

\item{path.to.prev}{a connection to the previsions for each sample and for each source group.}

\item{sample.name.obs, sample.name.prev}{a character string of the column name with sample name in obs/prev.}

\item{prev.source.name}{a character string of the column name with source groups in prev.}

\item{prev.values}{a character string of the column name with prevision values in prev.}

\item{source.groups}{a vector of character with source groups names.}

\item{mean.cal}{a boolean. Set TRUE to calculate the CRPS mean value per sample.}

\item{save.dir}{connection open for writing the test results data.frame. If "" save the file at working directory, if not set (default) the data.frame is not saved.}

\item{note}{a character string to add a note at the end of the file name (not set - default).}

\item{fileEncoding}{character string, if non-empty declares the encoding to be used on a file (not a connection) so the character data can be re-encoded
as they are written, "latin1" (default).}
}
\description{
Compute the CRPS values for Bayesian models.
}
\author{
Thomas Chalaux-Clergue
}
