% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_for_mixsiar.R
\name{data.for.MixSIAR}
\alias{data.for.MixSIAR}
\title{Generate CSV for MixSIAR}
\usage{
data.for.MixSIAR(
  data,
  class,
  target,
  tracers,
  sample.name,
  save.dir,
  note,
  fileEncoding = "latin1",
  show.data = FALSE
)
}
\arguments{
\item{data}{A data.frame containing all the data values.}

\item{class}{A character string corresponding to the column that contains the sources classes (source 1, source 2, etc) and target information.}

\item{target}{A character string corresponding to the way "Target" (default) are named.}

\item{tracers}{A vector listing the tracers names.}

\item{sample.name}{A character string correspond to samples id/name column}

\item{save.dir}{Connection open for writing the test results data.frame. If "" save the file at working directory.}

\item{note}{A character string to add a note at the end of the file name (not set - default).}

\item{fileEncoding}{character string, if non-empty declares the encoding to be used on a file (not a connection) so the character data can be re-encoded
as they are written, "latin1" (default).}

\item{show.data}{A boolean, if FALSE data.frame are saved but not returned.}
}
\description{
Generate mix and source CSV files compatible with MixSIAR data loading functions (i.e. *load_mix_data* and *load_source_data*).
}
\author{
Thomas Chalaux-Clergue
}
