% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_watcher.R
\name{data.watcher}
\alias{data.watcher}
\title{Evaluating measurement uncertainty}
\usage{
data.watcher(data, properties, prop.uncer, rl.lvl = 5)
}
\arguments{
\item{data}{a data.frame containing all the data values.}

\item{properties}{a character string corresponding to the column containing the sources (source 1, source 2, etc) and target information.}

\item{prop.uncer}{a vector listing all the properties that will be evaluated.}

\item{rl.lvl}{a numeric value indicating the relative uncertainty}
}
\value{
The list of properties with missing values, virtually impossibles and the size of the uncertainty compared to the property values.
}
\description{
Helps to consider the use of a property according to its measurement uncertainty.
}
\details{
Three types of errors are watched:
(a) missing values,
(b) positive property containing negative values,
(c) all negative values property (which require to be changed to positive for modelling),
(d) considering the uncertainty make some values virtually impossible (e.g. negative values within positive values property),
(e) relative size of the uncertainty compare to the value.
}
\author{
Thomas Chalaux-Clergue
}
