% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrim_tests.R
\name{discrim.tests}
\alias{discrim.tests}
\title{Discriminant assessement}
\usage{
discrim.tests(data, class, mixture, test, properties)
}
\arguments{
\item{data}{A data.frame containing at least class and property values.}

\item{class}{A character string corresponding to the column that contains the sources (source 1, source 2, etc) and mixture information if data contains mixture samples.}

\item{mixture}{A character string corresponding to way "mixture" are named in class. No name is defined by default. can not be set if date does not contain mixture samples.}

\item{test}{A character vector that indicates the test that will be performed: Kolmogorov-Smirnov ("KS") or Kruskal-Wallis ("KW").}

\item{properties}{A vector listing all the properties that will be evaluated.}
}
\value{
A data.frame with test results.
}
\description{
Run discriminant tests for every property. Test can be Kruskal-Wallis (KW) or Kolmogorov-Smirnov (KS).
}
\details{
This function is called within *fingR::discriminant.test*.
}
\author{
Thomas Chalaux-Clergue & Rémi Bizeul
}
