% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discriminant_power.R
\name{discriminant.power}
\alias{discriminant.power}
\title{Discriminant power}
\usage{
discriminant.power(data, test.context, min.discriminant = 1, p.level = 0.05)
}
\arguments{
\item{data}{A data.frame with all source group pairs results from tests.}

\item{test.context}{A character vector that indicates in which context two-samples Kolmogorov-Smirnov tests ("KS") were realised: Kolmogorov-Smirnov tests ("KS") or to test Kruskal-Wallis assumption ("KW.KS").}

\item{min.discriminant}{A numeric value of the minimum number of statistically different tests (1 default) for KS context.}

\item{p.level}{A numeric value of the p-value significance level.}
}
\description{
Identify the strength properties discriminant power based on two-samples tests.
}
\details{
Calculates how many source groups combinations exists (b) and how many of these combinations are statistically different based on the p-value (a).
The discriminant power can be assessed the ratio between a/b values.
This function is called within *fingR::discriminant.test* function.
}
\author{
Thomas Chalaux-Clergue & Rémi Bizeul
}
