% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discriminant_test.R
\name{discriminant.test}
\alias{discriminant.test}
\title{Discriminant power assessment}
\usage{
discriminant.test(
  data,
  class,
  mixture,
  properties,
  test,
  p.level = 0.05,
  min.discriminant = 1,
  save.discrim.tests = TRUE,
  save.dir,
  note,
  fileEncoding = "latin1"
)
}
\arguments{
\item{data}{A data.frame containing all the data values.}

\item{class}{A character string corresponding to the column that contains the sources classes (source 1, source 2, etc) and mixture information.}

\item{mixture}{A character string corresponding to way "mixture" (default) are named.}

\item{properties}{A vector listing all the properties that will be evaluated.}

\item{test}{A character vector that indicates the test that will be performed: Kolmogorov-Smirnov ("KS") or Kruskal-Wallis ("KW").}

\item{p.level}{p-value threshold to reject null hypothesis (H0) (0.05 - default).}

\item{min.discriminant}{A numeric value of the lowest number of sources couples that need to be statistically different to consider the property as discriminant (1 - default). Only for Kolmogorov-Smirnov ("KS").}

\item{save.discrim.tests}{Logical. If TRUE (default) save the data.frame of all sources pairs tests.}

\item{save.dir}{Connection open for writing the test results data.frame, "" save the file at working directory, if not set (default) the data.frame is not saved.}

\item{note}{A character string to add a note at the end of the file name (not set - default).}

\item{fileEncoding}{Character string, if non-empty declares the encoding to be used on a file (not a connection) so the character data can be re-encoded
as they are written, "latin1" (default).}
}
\value{
data.frame. The results of the different test.
}
\description{
Performs discrimination tests on data for fingerprinting studies.
Tests results highly depends on the defined sources classes.
}
\author{
Thomas Chalaux-Clergue & Rémi Bizeul
}
