% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensure_total.R
\name{ensure.total}
\alias{ensure.total}
\title{Ensure the total of contributions}
\usage{
ensure.total(
  data,
  sample.name,
  digits = 3,
  path,
  note,
  fileEncoding = "latin1"
)
}
\arguments{
\item{data}{A data.frame containing predicted contributions for each sources and sample names/id.}

\item{sample.name}{A character corresponding to the name of the column with sample names/id. If not set the function use the first column.}

\item{digits}{nteger indicating the number of decimal places (round) or significant digits (signif) to be used. For round, negative values are allowed (see base::Round for 'Details’).}
}
\value{
a data.frame.
}
\description{
Correct contributions to reach a total of 1 or 100%.
}
\author{
Thomas Chalaux-Clergue
}
