% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESP.R
\name{ESP}
\alias{ESP}
\title{Encompassed sample prediction}
\usage{
ESP(s1, s2)
}
\arguments{
\item{s1, s2}{Two vectors containing numerical values}
}
\value{
Numerical percentage of coverage.
}
\description{
Calculate the number of elements from s1 that lies within the s2 range.
}
\author{
Thomas Chalaux-Clergue
}
