% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esp_values.R
\name{esp.values}
\alias{esp.values}
\title{Encompassed sample predictions}
\usage{
esp.values(pred, obs, count = FALSE, digits = 0)
}
\arguments{
\item{obs, pred}{Two vectors containing numerical values.}

\item{count}{Boolean. Return percentage when FALSE (default) and count when TRUE.}

\item{digits}{Integer indicating the number of decimal places (round) or significant digits (signif) to be used. For round, negative values are allowed (see base::Round for more details).}
}
\description{
Calculate the number of elements from prediction (pred) that lies within the observed (obs) range.
}
\author{
Thomas Chalaux-Clergue
}
