% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_groups.R
\name{eval.groups}
\alias{eval.groups}
\title{Evaluation of source groups prediction quality}
\usage{
eval.groups(df.obs, df.pred, by, path, note, fileEncoding = "latin1")
}
\arguments{
\item{df.obs}{A data.frame with the observed source group contributions.}

\item{df.pred}{A data.frame with the predicted source group contributions.}

\item{by}{A character vector of variables to join by within the dplyr::left_join.}

\item{path}{Connection open for writing the test results data.frame. If "" save the file at working directory, if not set (default) the data.frame is not saved.}

\item{note}{A character string to add a note at the end of the file name (not set - default).}

\item{fileEncoding}{A character string, if non-empty declares the encoding to be used on a file (not a connection) so the character data can be re-encoded
as they are written, "latin1" (default).}
}
\description{
Calculate prediction quality statistics index to assess model prediction quality.
}
\details{
Call *fingR::eval.mix* function for the calculation of prediction statistics.
}
\author{
Thomas Chalaux-Clergue
}
