% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_mix.R
\name{eval.mix}
\alias{eval.mix}
\title{Evaluation of prediction quality}
\usage{
eval.mix(obs, pred)
}
\arguments{
\item{obs}{A numeric vector with observed values.}

\item{pred}{A numeric vector with predicted values}
}
\value{
A data.frame with
}
\description{
Calculates prediction quality statistic indexes: mean error (ME), root mean squared error (RMSE), Pearson correlation coefficient (r2) and
Nash-Stucliff modelling efficiency coefficient (NSE).
}
\references{
Bennett ND, Croke BFW, Guariso G et al (2013) Characterising perfor- mance of environmental models. Environ Model Softw 40:1–20. https:// doi. org/ 10. 1016/j. envso ft. 2012. 09. 011
}
\author{
Thomas Chalaux-Clergue and Rémi Bizeul
}
