% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gate_keeper.R
\name{gate.keeper}
\alias{gate.keeper}
\title{Verify and correct fully negative column}
\usage{
gate.keeper(data, tracers, inverter = FALSE)
}
\arguments{
\item{data}{A data.frame.}

\item{tracers}{A vector listing them columns.}

\item{inverter}{A boolean, if TRUE will invert the sign of designated columns}
}
\description{
Change fully negative column to fully positive.
}
\author{
Thomas Chalaux-Clergue
}
