% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPerms.R
\name{getPerms}
\alias{getPerms}
\title{Permutation calculator}
\usage{
getPerms(x)
}
\arguments{
\item{x}{A data.frame (ncols > 1) with numerir/character values.}
}
\value{
A data.frame with all the permutations row bound to x.
}
\description{
Generate all the permutations of the given data.frame.
}
\author{
Thomas Chalaux-Clergue
}
