% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_crps.R
\name{graph.crps}
\alias{graph.crps}
\title{Graphic representation of CRPS}
\usage{
graph.crps(
  obs,
  crps,
  by,
  path,
  note,
  units = c("cm"),
  width = 10,
  height = 10,
  line0 = TRUE
)
}
\arguments{
\item{obs}{A data.frame with observed values for each sample and each source group.}

\item{crps}{A data.frame with CRPS values for each sample and each source group.}

\item{by}{A character vector of variables to join by. Please refer to dplyr::mutate-joins vignette for more detailed explanations.}

\item{path}{Connection open for writing the test results data.frame, "" save the file at working directory, if not set (default) the data.frame is not saved.}

\item{note}{A character string to add a note at the end of the file name (not set - default).}

\item{width, height, units}{Plot size in units ("in", "cm", "mm", or "px"). If not supplied, uses the size of current graphics device.}
}
\description{
Draw source groups observed contributions by CRPS values.
}
\author{
Thomas Chalaux Clergue
}
