% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_op.R
\name{graph.op}
\alias{graph.op}
\title{Observed versus Predicted}
\usage{
graph.op(
  obs,
  pred,
  data,
  path,
  save.path,
  note,
  acceptance,
  labels = FALSE,
  text = FALSE,
  stats = FALSE,
  pt.colour,
  ln.colour,
  pt.pch = 1,
  units = c("cm"),
  width = 12,
  height = 12,
  format = c("pdf", "png"),
  dpi = 300
)
}
\arguments{
\item{obs}{A numeric vector with observed values. [not avalaible]}

\item{pred}{A numeric vector with predicted values. [not avalaible]}

\item{data}{A data.frame with observed and predicted values. [not avalaible]}

\item{path}{Connection to the folder were the ObsPred and stats files are located and connection open for writing the test results data.frame, "" save the file at working directory, if not set (default) the data.frame is not saved.}

\item{note}{A character string to add a note at the end of the file name (not set - default).}

\item{acceptance}{A numerical value to draw confidence interval around the 1:1 line.}

\item{labels}{A boolean, TRUE to display samples names with *geom_labels* function (FALSE, default).}

\item{text}{A boolean, TRUE to display samples names with *geom_text* function (FALSE, default).}

\item{stats}{A boolean, TRUE to display modelling quality statistics. (FALSE, default).}

\item{pt.pch}{Point shape aesthetics.}

\item{width, height, units}{Plot size in units ("in", "cm", "mm", or "px"). If not supplied, uses the size of current graphics device.}

\item{format, dpi}{saving format (pdf or png) and png dpi.}

\item{pt.coulour}{Point colour aesthetics.}

\item{ln.coulour}{Line colour aesthetics.}
}
\description{
Draw observed vers. predicted graphics.
}
\details{
ObsPred file is generated by *fingR::JAGS.summary* and contains: observed (obs), predicted values.
stats is generated with *fingR::eval.groups* function contains prediction quality statistics and value calculation Type (mean, median...) and source Group information.
}
\author{
Thomas Chalaux Clergue
}
