% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_conservative.R
\name{is.conservative}
\alias{is.conservative}
\title{Is conservative}
\usage{
is.conservative(
  data,
  property,
  test.format = "RT",
  crit.pos = 2,
  sep.format = "_",
  note
)
}
\arguments{
\item{data}{A data.frame containing all range test results.}

\item{property}{A character listing all the properties that will be evaluated.}

\item{test.format}{A character vector of format strings to recognise test ("test") and criteria ("crit") label position.}

\item{crit.pos}{A numeric value that indicates where criteria is wrote in test columns labels.}

\item{sep.format}{A character vector of format strings to recognise the separator to split test labels.}

\item{note}{A character string to add a note at the start of test label (not set - default).}
}
\value{
data.frame. The results of the different test.
}
\description{
Create vector(s) of discrimant properties according to their results
}
\author{
Thomas Chalaux-Clergue
}
