% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_discriminant.R
\name{is.discriminant}
\alias{is.discriminant}
\title{Is discriminant}
\usage{
is.discriminant(
  data,
  property,
  test.format,
  test.pos,
  sep.format,
  p.level = 0.05,
  note
)
}
\arguments{
\item{data}{A data.frame containing all range test results.}

\item{property}{A character listing all the properties that will be evaluated.}

\item{test.format}{A character vector of format strings to recognise test ("test") and criteria ("crit") label position.}

\item{test.pos}{A numeric value that indicates where criteria is wrote in test columns labels.}

\item{sep.format}{A character vector of format strings to recognise the separator to split test labels.}

\item{p.level}{p-value threshold to reject null hypothesis (H0) (0.05 - default).}

\item{note}{A character string to add a note at the start of test label (not set - default).}
}
\value{
data.frame. The results of the different test.
}
\description{
Create vector(s) of discrimant properties according to their results
}
\author{
Thomas Chalaux-Clergue
}
