% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JAGS_pred.R
\name{JAGS.pred}
\alias{JAGS.pred}
\title{JAGS Predictions}
\usage{
JAGS.pred(path, stats, save = FALSE, note, fileEncoding = "latin1")
}
\arguments{
\item{path}{a character string of the path to the prediction file generated by `fingR::JAGS.summary` function.}

\item{stats}{a character string indicating the statistic that is kept: "Median" or "Mean".}

\item{save}{a boolean to save predicted contributions}

\item{note}{a character string to add a note at the end of the file name (not set - default).}

\item{fileEncoding}{character string, if non-empty declares the encoding to be used on a file (not a connection) so the character data can be re-encoded as they are written, "latin1" (default).}
}
\description{
Generate and save a table with all the prediction of an JAGS model
}
\author{
Thomas Chalaux-Clergue
}
