% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JAGS_summary.R
\name{JAGS.summary}
\alias{JAGS.summary}
\title{Summary JAGS}
\usage{
JAGS.summary(
  jags.1,
  mix,
  sources,
  path,
  note,
  fileEncoding = "latin1",
  save_pred = FALSE
)
}
\arguments{
\item{jags.1}{rjags model object, output from MixSIAR::run_model function}

\item{mix}{output from MixSIAR::load_mix_data}

\item{sources}{output from MixSIAR::load_source_data}

\item{path}{connection open to write the results data.frame.}

\item{note}{a character string to add a note at the end of the file name (not set - default).}

\item{fileEncoding}{character string, if non-empty declares the encoding to be used on a file (not a connection) so the character data can be re-encoded
as they are written, "latin1" (default).}

\item{save_pred}{a boolean to save all the MixSIAR Monte-Carlo chain predictions}
}
\description{
Generate a table with each sample summary of JAGS predictions (e.g. mean, quantiles, etc).
}
\author{
Thomas Chalaux-Clergue
}
