% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvl_signif.R
\name{lvl.signif}
\alias{lvl.signif}
\title{Significance level}
\usage{
lvl.signif(x)
}
\arguments{
\item{x}{a numeric or a complex numeric vector.}
}
\description{
Return the significant figures level of a numeric or the highest level of a vector of numerics.
}
\author{
Thomas Chalaux-Clergue
}
