% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_tests.R
\name{range.tests}
\alias{range.tests}
\title{Range tests}
\usage{
\method{range}{tests}(
  data,
  class,
  mixture = "target",
  properties,
  alternative = "single",
  sample.id,
  criteria = "all",
  MM.error = c(0.1),
  hinge.range = c(70),
  CI.lvl = c(95),
  shift = 0.1,
  save.each = TRUE,
  save.dir,
  note,
  fileEncoding = "latin1"
)
}
\arguments{
\item{data}{A data.frame containing all the data values.}

\item{class}{A character string corresponding to the column that contains the sources classes (source 1, source 2, etc) and mixture information.}

\item{mixture}{A character string corresponding to way "mixture" (default) are named.}

\item{properties}{A vector listing all the properties that will be evaluated.}

\item{alternative}{A character that indicates if range test should be apply on each sample (single) or should calculates the criterion on all samples (population).}

\item{sample.id}{description}

\item{criteria}{A vector or character that indicates the range tests criteria to be performed among: minimum-maximum ("MM"), minimum-maximum +/- measurement error ("MMe"), "whiskers", "hinge", "mean", mean +/- standard deviation ("mean.sd"), "median".}

\item{MM.error}{A numeric value indicating how much measurement error should be added to source value. No measurement error = 0.0 and 10 percent = 0.1.}

\item{hinge.range}{A vector listing tested percentages of population within the hinge, 70 -> c(70) (default).}

\item{CI.lvl}{A vector listing tested confidance interval level, 95 = c(95) (default).}

\item{shift}{A numeric value of the size of shift to correct the minimum value to do log transformation, .1 (default)}

\item{save.each}{TRUE (default) to save each sample range test.}

\item{save.dir}{Connection open for writing the test results data.frame. If "" save the file at working directory. If not set (default) the data.frame is not saved.}

\item{note}{A character string to add a note at the end of the file name (not set - default).}

\item{fileEncoding}{Character string, if non-empty declares the encoding to be used on a file (not a connection) so the character data can be re-encoded
as they are written, "latin1" (default).}
}
\value{
a data.frame with tests results.
}
\description{
Performs range tests on properties to assess conservative behaviour for sediment fingerprinting studies.
}
\author{
Thomas Chalaux-Clergue & Rémi Bizeul
}
