% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RT_hinge.R
\name{RT.hinge}
\alias{RT.hinge}
\title{Hinge range test}
\usage{
RT.hinge(target, sources, class, h.range, alternative = "single")
}
\arguments{
\item{target}{a (non-empty) numeric vector of target values.}

\item{sources}{a (non-empty) data.frame with information of source classes and property value.}

\item{class}{A character string corresponding to the column that contains the sources classes (source 1, source 2, etc) and target information.}

\item{h.range}{A numeric value of the population percentage within the hinge}

\item{alternative}{A character that indicates if range test should be apply on each sample (sample) or should calculates the criterion on all samples (population).}
}
\description{
Performs a range test based on the hinge/interstats::quantile range (IQR) criterion.
}
\author{
Thomas Chalaux-Clergue & Rémi Bizeul
}
