% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RT_median.R
\name{RT.median}
\alias{RT.median}
\title{Median criterion range test}
\usage{
RT.median(target, sources, class, alternative = "single")
}
\arguments{
\item{target}{a (non-empty) numeric vector of target values.}

\item{sources}{a (non-empty) numeric vector of source values.}

\item{class}{name of the source classes columns in sources.}

\item{alternative}{a character that indicates if range test should be apply on each sample (sample) or should calculates the criterion on all samples (population).}
}
\description{
Assess the conservativity of a property based on the comparison of the sources and target medians.
The property is conservative if the target median value is between the highest and lowest sources medians.
}
\author{
Thomas Chalaux-Clergue & Rémi Bizeul
}
