% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RT_MM.R
\name{RT.MM}
\alias{RT.MM}
\title{Minimum-maximum criterion range test}
\usage{
RT.MM(target, sources, class, MM.error = 0, alternative = "single")
}
\arguments{
\item{target}{A (non-empty) numeric vector of target values.}

\item{sources}{A (non-empty) data.frame with information of source classes and property value.}

\item{class}{A character string corresponding to the column that contains the sources classes (source 1, source 2, etc) and target information.}

\item{MM.error}{A numeric value indicating how much measurement error should be added to source value. No measurement error = 0.0 and 10 percent = 0.1.}

\item{alternative}{A character that indicates if range test should be apply on each sample (sample) or should calculates the criterion on all samples (population).}
}
\description{
Performs a range test based on the minimum-maximum criterion. For a single target, it should be above the lowest minimum and bellow the highest maximum of source groups. For a population, the minimum target sample should be above the lowest minimum value of source groups and the maximum target sample should be bellow the highest maximum of source groups.
}
\author{
Thomas Chalaux-Clergue & Rémi Bizeul
}
