% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RT_player.R
\name{RT.player}
\alias{RT.player}
\title{Run range criteria}
\usage{
RT.player(
  data.t,
  data.s,
  criteria,
  MM.error,
  hinge.range,
  CI.CLs,
  alternative = "single",
  class,
  shift = 0.1
)
}
\arguments{
\item{data.t}{A numeric vector with source property value}

\item{data.s}{A numeric vector with source property value or if alternative is set to "independent" a matrix with sources class and property value (2 columns).}

\item{criteria}{Indicates the test(s) that will be performed, "all" or a combination between: "MM", "hinge", "whiskers", "CI" (only merged), "OSDM" (only merged), "mean" (only independent), or "median" (only merged).}

\item{MM.error}{A numeric value indicating how much measurement error should be added to source value. No measurement error = 0.0 and 10 percent = 0.1.}

\item{hinge.range}{A vector listing tested percentages of population within the hinge, 70 -> c(70) (default).}

\item{CI.CLs}{A vector listing tested confidance interval level, 95 = c(95) (default).}

\item{alternative}{A character that indicates if range test should be apply on each sample (sample) or should calculates the criterion on all samples (population).}

\item{class}{A character string corresponding to the column that contains the sources classes (source 1, source 2, etc) and target information.}

\item{shift}{A numeric value of the size of shift to correct the minimum value to do log transformation, .1 (default)}
}
\description{
This function is called within *is.conservative* and was not designed for human use.
}
\author{
Thomas Chalaux-Clergue & Rémi Bizeul
}
