% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RT_whiskers.R
\name{RT.whiskers}
\alias{RT.whiskers}
\title{Whiskers criterion range test}
\usage{
RT.whiskers(target, sources, class, alternative = "single")
}
\arguments{
\item{target}{a (non-empty) numeric vector of target values.}

\item{sources}{a (non-empty) data.frame with information of source classes and property value.}

\item{class}{a character string corresponding to the column that contains the sources classes (source 1, source 2, etc) and target information.}

\item{alternative}{a character that indicates if range test should be apply on each sample (single) or should calculates the criterion on all samples (population).}
}
\description{
Performs a range test based on the whiskers criterion.
}
\author{
Thomas Chalaux-Clergue & Rémi Bizeul
}
