% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_BMM.R
\name{run.BMM}
\alias{run.BMM}
\title{run.BMM}
\usage{
run.BMM(
  data,
  class,
  mixture = "target",
  sample.id,
  tracers,
  uncertainty,
  isotope.ratio,
  isotope.prop,
  isotopes.unc,
  save.dir,
  note,
  fileEncoding = "latin1",
  multivar = TRUE,
  sd.threshold = 0.05,
  sample.size = 100,
  n.normal.targets = 2500,
  n.iter = 2500,
  lower.bound = 0.001,
  upper.bound = 1,
  constraints = 1
)
}
\arguments{
\item{data}{Data frame. The dataset containing both source and target samples.}

\item{class}{Character. The column name that specifies the class (source/target).}

\item{mixture}{Character. The class label for the target samples (default is \code{"target"}).}

\item{sample.id}{Character. The column name for sample identifiers.}

\item{tracers}{Character vector. The column names of the tracers used in the model.}

\item{uncertainty}{Character vector. Optional column names for uncertainty values of the tracers.}

\item{isotope.ratio}{Character vector. Optional column names for the property isotopes ratio (e.g. d13C).}

\item{isotope.prop}{Character vector. Optional column names for the concentration properties of the respective isotopes (e.g. TOC).}

\item{isotopes.unc}{Character vector. Optional column names for isotopes uncertainties (e.g. d13C_uncertainty).}

\item{save.dir}{Character. Optional directory path for saving the results.}

\item{note}{Character. Optional additional note to append to the file name.}

\item{fileEncoding}{Character. Encoding to be used for the saved file (default is \code{"latin1"}).}

\item{multivar}{Logical. Whether to use multivariate normal distribution for sources (default is \code{TRUE}).}

\item{sd.threshold}{Numeric. Standard deviation threshold for generating target samples (default is 0.05).}

\item{sample.size}{Integer. The size of the sample to generate for the target (default is 100).}

\item{n.normal.targets}{Integer. The number of normally distributed samples for sources (default is 2500).}

\item{n.iter}{Integer. The number of iterations for the model simulation (default is 2500).}

\item{lower.bound}{Numeric. The lower bound for the source contributions (default is 0.001).}

\item{upper.bound}{Numeric. The upper bound for the source contributions (default is 1).}

\item{constraints}{Numeric. The constraints for the equality function in the model (default is 1).}
}
\value{
A data frame with the estimated source contributions for each target sample.
}
\description{
Runs the Bayesian Mixing Model (BMM) to estimate source contributions for given target samples. Source contribution are estimated by minimising the sum of squared residuals (SSR) of the mass balance un-mixing model.
Manage isotopic ratio for tracers specified within \code{isotope.prop}, \code{isotope.ratio} (optional \code{isotopes.unc})
}
\author{
Thomas Chalaux-Clergue & Rémi Bizeul
}
