% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selected_tracers.R
\name{selected.tracers}
\alias{selected.tracers}
\title{Selected tracers}
\usage{
selected.tracers(cons, discrim, note)
}
\arguments{
\item{cons}{a vector or a list of vector that contain(s) conservatives properties}

\item{discrim}{a vector or a list of vector that contain(s) discriminant properties}

\item{note}{A character string to add a note at the start of test label (not set - default).}
}
\value{
data.frame. The results of the different test.
}
\description{
Create a list of vector(s) containing conversative and discriminant properties.
}
\author{
Thomas Chalaux-Clergue
}
