% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_norm_distrib.R
\name{source.norm.distrib}
\alias{source.norm.distrib}
\title{Normal distributed source samples}
\usage{
source.norm.distrib(data, class, tracers, n = 2500, multivar = FALSE)
}
\arguments{
\item{data}{a data.frame containing all the source properties values.}

\item{class}{a character string corresponding to the column that contains the sources classes (source 1, source 2, etc) and target information.}

\item{tracers}{a vector listing all the tracers column names.}

\item{n}{a numeric value, the number of samples generated.}

\item{multivar}{a boolean, TRUE to generate multivariate normal distribution of source properties (FALSE, default).}
}
\description{
Generate artificial (multivariate) normal distributed source samples.
}
\references{
Batista, P. V. G., Laceby, J. P., & Evrard, O. (2022). How to evaluate sediment fingerprinting source apportionments. Journal of Soils and Sediments, 22(4), 1315-1328.
}
\author{
Thomas Chalaux-Clergue
}
