% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepwise_selection.R
\name{stepwise.selection}
\alias{stepwise.selection}
\title{A Wilk's lambda stepwise tracer selection}
\usage{
stepwise.selection(data, class, tracers, target, save.dir, note)
}
\arguments{
\item{data}{a data.frame containing all the data values.}

\item{class}{a character string corresponding to the column that contains the sources (source 1, source 2, etc) and target information.}

\item{tracers}{a vector listing all the tracers that will be evaluated.}

\item{target}{a character string corresponding to way "Target" are named, not set default.}

\item{save.dir}{connection open for writing the test results data.frame, "" save the file at working directory, if not set (default) the data.frame is not saved.}

\item{note}{a character string to add a note at the end of the file name (not set - default).}
}
\value{
a vector of selected tracers
}
\description{
Run a discriminant function analysis (DFA) forward stepwise selection based on Wilk's criterion to select tracers that maximise the difference and minimise the redoundancy of information among selected tracers.
}
\author{
Thomas Chalaux-Clergue and Rémi Bizeul
}
