% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_log.R
\name{to.log}
\alias{to.log}
\title{Log transformation}
\usage{
to.log(data, shift = 0.1)
}
\arguments{
\item{data}{a (non-empty) numeric vector of data .}

\item{shift}{a numeric value of the size of shift, .1 (default)}
}
\description{
Convert data to log, while being careful to negatives values.
If some values are negative performs a shift of all values to the minimum value increased by a certain percentage,  10% (.1 - default).
}
\author{
Thomas Chalaux-Clergue
}
