% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VM_builder.R
\name{VM.builder}
\alias{VM.builder}
\title{Virtual mixtures property values generator}
\usage{
VM.builder(
  data,
  material,
  source.name = "Source",
  class,
  tracers,
  contributions,
  VM.name,
  Normal.distrib.samples = FALSE,
  multivar = FALSE,
  n.norm = 2500,
  step = 0.1,
  save.dir,
  note,
  fileEncoding = "latin1",
  RETURN = TRUE
)
}
\arguments{
\item{data}{A data.frame containing all the source properties values, material and class.}

\item{material}{A character string corresponding to the column where the source samples can be differentiate from other samples.}

\item{source.name}{A character string corresponding to the way source samples are named in 'material': "Source" (default).}

\item{class}{A character string corresponding to the column that contains the sources classification groups (source A, source B...) and target.}

\item{tracers}{A vector listing all the tracers column names.}

\item{contributions}{A data.frame with the artificial mixture source groups contributions (from 0 to 1). If not set proportions will be generated according to the 'step'.}

\item{VM.name}{A character string corresponding to the column where artificial mixture are labelled.}

\item{Normal.distrib.samples}{description}

\item{multivar}{A boolean, TRUE to generate multivariate normal distribution of source properties (FALSE, default).}

\item{n.norm}{A numeric value corresponding the number of samples generated.}

\item{step}{A numeric value indicating the step between two contributions, between 0 and 1 (default: 0.1).}

\item{save.dir}{Connection open for saving artificial mixtures property values and standard deviation (SD) data.frames. If "" save the file at working directory, if not set (default) the data.frame is not saved.}

\item{note}{A character string to add a note at the end of the file name (not set - default).}

\item{RETURN}{A boolean, when set to TRUE (default) return the artificial mixture property values. Set to FALSE for no return from the function.}

\item{fieEncoding}{Character string, if non-empty declares the encoding to be used on a file (not a connection) so the character data can be re-encoded
as they are written, "latin1" (default).}
}
\value{
A list of two data.frame: artificial mixtures *property* values and artificial mixtures properties measurement *uncertainty*.
}
\description{
Generate artificial mixture property values according to source samples properties.
}
\details{
Virtual mixture properties values are determined according to the multiplication of each source distribution (%) and a randomly selected value among the Normal distributed
samples (n = n.norm). Normal distributed samples are generated according to each source group characteristics (mean and standard deviation) within the *fingR::source.norm.distrib* function.
}
\references{
Batista, P. V. G., Laceby, J. P., & Evrard, O. (2022). How to evaluate sediment fingerprinting source apportionments. Journal of Soils and Sediments, 22(4), 1315-1328.
}
\author{
Thomas Chalaux-Clergue
}
