% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VM_contrib_generator.R
\name{VM.contrib.generator}
\alias{VM.contrib.generator}
\title{Virtual mixtures proportion generator}
\usage{
VM.contrib.generator(
  n.sources,
  groups,
  step = 0.1,
  save.dir,
  note,
  fileEncoding = "latin1"
)
}
\arguments{
\item{n.sources}{A numeric value indicating the number of source groups.}

\item{groups}{A vector listing source groups names}

\item{step}{A numeric value indicating the step between two contributions, between 0 and 1 (default: 0.1).}

\item{save.dir}{A connection open for saving artificial mixtures property values and standard deviation (SD) data.frames. If "" save the file at working directory, if not set (default) the data.frame is not saved.}

\item{note}{A character string to add a note at the end of the file name (not set - default).}

\item{fileEncoding}{Character string, if non-empty declares the encoding to be used on a file (not a connection) so the character data can be re-encoded
as they are written, "latin1" (default).}
}
\value{
A data.frame with artificial mixture source contributions.
}
\description{
Generate proportion for artificial mixture according to the number of source groups, from 2 to 5.
}
\details{
Contributions are generated according to a scale range from 0 to 100% contribution for a source, with the *step* increments.
The other groups are determined as fractions of the remaining contribution (i.e. 1 - source A contribution) with decreasing importance of groups.
The value of the fraction is determined by the number of sources. The denominator is defined as: (number of sources - 1) * 2.
The numerator values are set according to the number groups: 3 groups - 3 and 1, 4 groups - 3, 2 and 1, 5 groups - 4, 3, 2 and 1.
Then, the permutation of this contribution scale are determine.
}
\author{
Thomas Chalaux-Clergue
}
