% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VM_normal_distrib_prop_values.R
\name{VM.normal.distrib.prop.values}
\alias{VM.normal.distrib.prop.values}
\title{Virtual mixtures normal distributed property values}
\usage{
VM.normal.distrib.prop.values(
  data,
  class,
  tracers,
  n,
  multivar,
  contributions,
  VM.name
)
}
\arguments{
\item{data}{description}

\item{class}{description}

\item{tracers}{description}

\item{n}{description}

\item{multivar}{description}

\item{contributions}{description}

\item{VM.name}{description}
}
\value{
A list of two data.frame: artificial mixtures *property* values and artificial mixtures properties measurement *uncertainty*.
}
\description{
Generate virtual mixtures property values using a set of normal distributed samples relying of source group property mean and standard deviation.
}
\references{
Batista, P. V. G., Laceby, J. P., & Evrard, O. (2022). How to evaluate sediment fingerprinting source apportionments. Journal of Soils and Sediments, 22(4), 1315-1328.
}
\author{
Thomas Chalaux-Clergue
}
