% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_disc_tests.R
\name{which.disc.tests}
\alias{which.disc.tests}
\title{Discriminant test label}
\usage{
which.disc.tests(tests)
}
\arguments{
\item{tests}{a non-empty character vector of tests names: Kolmogorov-Smirnov (KW - default) or Kolmogorov-Smirnov ("KS").}
}
\description{
Return columns names for discriminant tests result data frame base on the setted list of tests.
This function is called by discriminant_power(). Not intended for user use.
}
\author{
Thomas Chalaux-Clergue
}
