% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinations.R
\name{combinations}
\alias{combinations}
\title{Combination generator}
\usage{
combinations(x)
}
\arguments{
\item{x}{A character vector containing elements to combine.}
}
\value{
A character vector containing combinations of elements in 'x' separated by "-".
}
\description{
Generate combinations of elements in a vector
}
\author{
Thomas Chalaux-Clergue
}
