% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESP.R
\name{ESP}
\alias{ESP}
\title{Encompassed sample predictions}
\usage{
ESP(
  obs,
  pred,
  sources,
  sources.obs,
  sources.pred,
  count = FALSE,
  digits = 0,
  split = "_"
)
}
\arguments{
\item{obs, pred}{Two data.frame containing numerical values.}

\item{sources, source.obs, sources.pred}{Character or vector of character indicating the column(s) that contains source labels. If source labels differ in obs and pred, indicate them.}

\item{count}{Boolean or Character. Return percentage when FALSE (default) and count when TRUE. Or "Both" if user want both.}

\item{digits}{Integer indicating the number of decimal places (round) or significant digits (signif) to be used. For round, negative values are allowed (see base::Round for more details).}

\item{split}{Character string used to split the row names of the result to extract the actual source names (e.g., if row names are "ESP.Percentage_SourceA", and split = "_", "SourceA" will be extracted). Default is "_".}
}
\description{
Calculate the number of elements from prediction (pred) that lies within the observed (obs) range for each source
}
\author{
Thomas Chalaux-Clergue
}
