% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval_width.R
\name{interval.width}
\alias{interval.width}
\title{Interval width}
\usage{
interval.width(
  prev,
  path.to.prev,
  sample_name,
  source.groups,
  value,
  mean.cal = FALSE,
  save = FALSE,
  save.dir,
  note,
  fileEncoding = "latin1"
)
}
\arguments{
\item{prev}{description}

\item{path.to.prev}{description}

\item{sample_name}{description}

\item{source.groups}{description}

\item{value}{description}

\item{mean.cal}{description}

\item{save}{description}

\item{note}{a character string to add a note at the end of the file name (not set - default).}

\item{fileEncoding}{character string, if non-empty declares the encoding to be used on a file (not a connection) so the character data can be re-encoded
as they are written, "latin1" (default).}

\item{save.path}{description}
}
\description{
Interval width
}
\author{
Thomas Chalaux-Clergue
}
