% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VM_contrib_generator.R
\name{VM.contrib.generator}
\alias{VM.contrib.generator}
\title{VM.contrib.generator}
\usage{
VM.contrib.generator(
  n.sources,
  min = 0,
  max = 100,
  step = 5,
  sources.class,
  VM.name = "Sample_name",
  save.dir,
  note,
  fileEncoding = "latin1",
  return = TRUE,
  save = TRUE
)
}
\arguments{
\item{n.sources}{Integer. The number of source levels (from 2 to 4).}

\item{min}{Numeric. Minimum contribution value (default is 0).}

\item{max}{Numeric. Maximum contribution value (default is 100).}

\item{step}{Numeric. Step between two contribution levels (default is 5).}

\item{sources.class}{Character vector. Optional names for the source classes.}

\item{VM.name}{Character. Name of the column containing virtual mixture labels (default is "Sample_name").}

\item{save.dir}{Connection open for saving artificial mixtures property values and standard deviation (SD) data.frames. If "" save the file at working directory, if not set (default) the data.frame is not saved.}

\item{note}{Character. Optional additional note to append to the file name.}

\item{return}{Logical. Whether to return the result matrix (Default is TRUE).}

\item{save}{Logical. Whether to save the result matrix to a file (Default is TRUE).}

\item{fieEncoding}{Character string, if non-empty declares the encoding to be used on a file (not a connection) so the character data can be re-encoded
as they are written (default is "latin1").}
}
\value{
A matrix (or data frame) of virtual mixture contributions if \code{return} is set to TRUE.
}
\description{
Generates a matrix of virtual mixture source contributions based on specified parameters.
}
\author{
Thomas Chalaux-Clergue
}
