% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VM_proportionate_prop_values.R
\name{VM.proportionate.prop.values}
\alias{VM.proportionate.prop.values}
\title{Virtual mixtures proportionate property values}
\usage{
VM.proportionate.prop.values(data, class, tracers, contributions, VM.name)
}
\arguments{
\item{data}{description}

\item{class}{description}

\item{tracers}{description}

\item{contributions}{description}

\item{VM.name}{description}
}
\value{
A list of two data.frame: artificial mixtures *property* values and artificial mixtures properties measurement *uncertainty*.
}
\description{
Generate virtual mixtures property values using source group property mean and standard deviation.
}
\references{
Batista, P. V. G., Laceby, J. P., & Evrard, O. (2022). How to evaluate sediment fingerprinting source apportionments. Journal of Soils and Sediments, 22(4), 1315-1328.
}
\author{
Thomas Chalaux-Clergue
}
